+++ src/game/game_math_functions.c
@@ -63,7 +63,7 @@ t_new_xy Intersect(float x1, float y1, float x2, float y2,float x3, float y3, fl
     return (xy);
 }
 
-int	Intersect_divider(float x1, float y1, float x2, float y2,float x3, float y3, float x4, float y4)
+float	Intersect_divider(float x1, float y1, float x2, float y2,float x3, float y3, float x4, float y4)
 {
 	return (vxs((x1)-(x2), (y1)-(y2), (x3)-(x4), (y3)-(y4)));
 }
